ch6/6.4/association/requirements.txt
tornado
neo4j-driver==1.4
marisa-trie
loguru
gensim